added view branch and compare branch button
